fix: resolve fileExporter generic inference and LocationExporter scope errors
- Move LocationExporter enum into SettingsView.swift to avoid xcodeproj
  registration issues with newly added files
- Replace optional csvFile state with exportData: Data so CSVFile is
  constructed inline, giving the compiler an unambiguous concrete type
  for the fileExporter generic parameter D
